Feature/external control link rotations (#40)
The primary goal of this commit is to remove all usages of "vertex bag states", and instead build magnetic time-evolution circuits by using a multiplicity index associated with each vertex along with the control links attached to the vertex. It's a kind of big PR (30 files changes), but it's not quite as big as it seems at first glance (20 of the file changes are adding/removing json data, 10 are actual codebase changes). There is a detailed explanation of the changes per-file in the PR.